move parser types to eolearn.core.types
