+++ features/eolearn/features/blob.py
@@ -16,7 +16,7 @@ import numpy as np
 import skimage.feature
 
 from eolearn.core import EOPatch, EOTask
-from eolearn.core.utils.parsing import SingleFeatureSpec
+from eolearn.core.types import SingleFeatureSpec
 
 
 class BlobTask(EOTask):
