+++ features/eolearn/features/clustering.py
@@ -16,8 +16,7 @@ from sklearn.cluster import AgglomerativeClustering
 from sklearn.feature_extraction.image import grid_to_graph
 
 from eolearn.core import EOPatch, EOTask, FeatureType
-from eolearn.core.types import Literal
-from eolearn.core.utils.parsing import FeatureSpec
+from eolearn.core.types import FeatureSpec, Literal
 
 
 class ClusteringTask(EOTask):
