+++ features/eolearn/features/local_binary_pattern.py
@@ -13,7 +13,7 @@ import numpy as np
 import skimage.feature
 
 from eolearn.core import EOPatch, EOTask
-from eolearn.core.utils.parsing import SingleFeatureSpec
+from eolearn.core.types import SingleFeatureSpec
 
 
 class LocalBinaryPatternTask(EOTask):
